requirements.txt
# Requisitos para executar o AsimoBot com LangChain e Groq
# Compatível com Python 3.12.3

langchain==0.1.20
langchain-groq==0.0.3
langchain-community==0.0.36
youtube-transcript-api==0.6.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0

# Comando para instalar todas as dependências acima:
# pip install -r requirements.txt
